Move Netcoreapp specific stuff to netcorapp file
